+++ Source/BugsnagConfiguration.m
@@ -134,6 +134,7 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     [copy setPlugins:[self.plugins copy]];
     [copy setReleaseStage:self.releaseStage];
     [copy setSession:[self.session copy]];
+    [copy setSendThreads:self.sendThreads];
     [copy setUser:self.user.userId
         withEmail:self.user.emailAddress
           andName:self.user.name];
@@ -196,6 +197,7 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     _redactedKeys = @[@"password"];
     _breadcrumbs = [BugsnagBreadcrumbs new];
     _autoTrackSessions = YES;
+    _sendThreads = BSGThreadSendPolicyAlways;
     // Default to recording all error types
     _enabledErrorTypes = BSGErrorTypesCPP
                        | BSGErrorTypesMach